Update the program
