+++ Cargo.toml
@@ -10,7 +10,7 @@ crate-type = ["cdylib"]
 [dependencies]
 gstd = { git = "https://github.com/gear-tech/gear.git", features = ["debug"] }
 
-codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive", "full"] }
+codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive"] }
 hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
 scale-info = { version = "1.0.0", default-features = false, features = ["derive"] }
 circular_buffer = {git = "https://github.com/breathx/circular-buffer-rs.git", default-features = false }
@@ -18,4 +18,4 @@ primitive-types = { version = "0.10.1", default-features = false, features = ["s
 
 [profile.release]
 lto = true
-opt-level = 's'
+opt-level = "s"
